project/requirements.txt
datasets==2.20.0
hydra-core==1.3.2
mlflow==2.14.1
numpy==1.23.5
torch==2.3.1
transformers==4.41.2